requirements.txt

streamlit
requests
python-whois
